requirements.txt
python-telegram-bot==20.7
anthropic==0.25.0
python-dotenv==1.0.0
aiohttp==3.9.3
aiosqlite==0.20.0
asyncio==3.4.3
tavily-python==0.3.3
rich==13.7.0
